Add some comments - Move header & add github buttons
