+++ src/App.tsx
@@ -1,55 +1,23 @@
 import React from 'react';
-import logo from './img/logo.svg';
 import './styles/app.css';
-import { Route, Routes, BrowserRouter, Link } from 'react-router-dom';
+import { Route, Routes, BrowserRouter } from 'react-router-dom';
 import HomePage from './routes/HomePage';
 import Login from './routes/Login';
 import Signup from './routes/Signup';
 import Quizzes from './routes/Quizzes';
 import AddEditQuiz from './routes/AddEditQuiz';
-import { Button, Header } from 'semantic-ui-react';
-import { getSignedInUserName } from './helpers/user';
-import Cookies from 'js-cookie';
 import PlayQuiz from './routes/PlayQuiz';
 import { useAppStore } from './useAppStore';
 import ConfirmationModal from './components/ConfirmationModal';
+import Header from './components/Header';
 
 function App() {
-  const userName = getSignedInUserName();
   const { confirmationModal } = useAppStore();
 
   return (
     <div className="App">
       <BrowserRouter>
-        <Header className="App-header">
-          <div className="flex alignCenter">
-            <Link to="/">
-              <img src={logo} className="App-logo" alt="logo" />
-            </Link>
-            {userName ? (
-              <>
-                <div className="mx-lg">{userName}</div>
-                <Button
-                  color="brown"
-                  onClick={() => {
-                    Cookies.set('sessionId', '', {
-                      domain: window.location.hostname,
-                      sameSite: 'Strict',
-                    });
-                    Cookies.set('userName', '', {
-                      domain: window.location.hostname,
-                      sameSite: 'Strict',
-                    });
-                    window.location.href = '/';
-                  }}>
-                  Logout
-                </Button>
-              </>
-            ) : (
-              <div className="mx-lg">Quizmaster</div>
-            )}
-          </div>
-        </Header>
+        <Header />
         <Routes>
           <Route path="/" element={<HomePage />} />
           <Route path="/login" element={<Login />} />
